pyproject.toml
[project]
name = "langlens"
version = "0.1.1"
description = "A visual tracer and debugger for LangChain and LangGraph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-core",
]

[project.urls]
Homepage = "https://github.com/viet-the-le/langlens"
Repository = "https://github.com/viet-the-le/langlens"
Issues = "https://github.com/viet-the-le/langlens/issues"

[tool.setuptools]
packages = ["langlens", "langlens.viewer"]
include-package-data = true

[tool.setuptools.package-data]
"langlens.viewer" = ["static/**/*"]

[project.optional-dependencies]
demo = [
    "langchain[google-genai]>=1.2.2",
    "langchain-ollama>=1.0.1",
    "langgraph-cli[inmem]>=0.4.11",
    "langchain-groq>=1.1.1",
    "python-dotenv",
]

[project.scripts]
langlens = "langlens.cli:main"
